refactor: move mock_term to separate file
